renaming(mkDocs): To mkdocs
Suggested by kmatyukhin on https://github.com/backstage/backstage/pull/15417
