+++ plugins/techdocs-node/api-report.md
@@ -88,6 +88,20 @@ export const getLocationForEntity: (
   scmIntegration: ScmIntegrationRegistry,
 ) => ParsedLocationAnnotation;
 
+// @public @deprecated (undocumented)
+export const getMkDocsYml: (
+  inputDir: string,
+  siteOptions?:
+    | {
+        name?: string | undefined;
+      }
+    | undefined,
+) => Promise<{
+  path: string;
+  content: string;
+  configIsTemporary: boolean;
+}>;
+
 // @public
 export const getMkdocsYml: (
   inputDir: string,
