+++ plugins/techdocs-node/src/stages/generate/helpers.test.ts
@@ -34,7 +34,7 @@ import {
 import {
   patchMkdocsYmlPreBuild,
   pathMkdocsYmlWithTechdocsPlugin,
-} from './mkDocsPatchers';
+} from './mkdocsPatchers';
 import yaml from 'js-yaml';
 
 const mockEntity = {
