+++ plugins/techdocs-node/src/stages/generate/mkdocsPatchers.ts
